Swap the ordinals in Hash.
...while we still can, on account of having no users to break.

We'll be using sha256 for the foreseeable future, so while it's good to
have the algo field in there fore easy upgradinging in the future, it's
a shame to have it take up extra space on the wire. By putting it
second, it will be erased during canonicalization (unless it is set to
something else).