Get rid of custom setup.
A couple problems with it:

- I never managed to debug the change detection, so it was rebuilding
everything every time. I was working around this by basically having
most of Setup.hs commented out in my working tree anyway.
- It's reasonably common for me to want to look at the generated code
to remind myself what the API is, and that's annoying when it's deep
within gensrc/

Instead, we now have to manually run ./gen-schema.sh. Better solutions
probably exist, but this is good for now.