Wire up get to the cli
Some bugs discovered by testing, which need to be fixed:

1. If you pass a path like "src/" during put, you get a file called "",
which is an illegal name that bombs out when doing a get.
2. Checking for a symbolic link throws an exception if the file isn't
there at all.

We should also probably rework the data types, so that a file's name is
in its dirent, not the rest of the metadata; otherwise there's no
obvious way to upload e.g. '/', since that path doesn't have a name...