Try to test more.
Prior to this patch, the coverage report reveals two things that aren't
tested:

1. Some error returns, most of these cascades
2. The case in savePair() where k1[0] == k2[0]

This patch attempts to test (2) by re-ordering statements to trigger it.
But for some reason, the Insert() call is returning ErrNotFound. I am
very confused, as I don't think it should be possible for Insert to
return this error under any circumstances.

The patch also adds some context to errors when they are reported by
chkfatal, useful for debugging.