Solve the null client issue, remove an indirection.
The problem we were seeing was because we were setting the file's root
blobtree ref to the ref for the blobtree's *contents*, rather than the
blobtree itself.

Per the now deleted TODO, the indirection is not really desirable in the
first place, so this patch:

1. Removes the Ref wrapper around File's blobtree field.
2. encodes the root directly into that field, rather than uploading it
   as its own ref.

The file we subsequently download is not identical to the one we upload
though, so more investigation is needed.